backend/requirements.txt
tensorflow
numpy
flask
Pillow
flask-cors
gunicorn
requests